KnowWE no longer allows users to build or change an article with &version=x or &parse=full without write permissions to that article.

+++ KnowWE/KnowWE-Plugin-JSPWiki-Connector/src/main/java/de/knowwe/jspwiki/KnowWEPlugin.java
@@ -156,6 +156,10 @@ public class KnowWEPlugin extends BasicPageFilter implements WikiPlugin,
 		try {
 			updateArticle(wikiContext, content);
 		}
+		catch (UpdateNotAllowedException e) {
+			String title = wikiContext.getPage().getName();
+			Log.fine("Somebody tried to update article " + title + " without appropriate rights", e);
+		}
 		catch (Exception e) {
 			String title = wikiContext.getPage().getName();
 			Log.severe("Exception while compiling article " + title, e);
@@ -265,13 +269,17 @@ public class KnowWEPlugin extends BasicPageFilter implements WikiPlugin,
 			userContext.getRequest().setAttribute("renderresult", stringRaw);
 			return stringRaw;
 		}
+		catch (UpdateNotAllowedException e) {
+			Log.fine("Somebody tried to update article " + title + " without appropriate rights", e);
+			return getExceptionRendering(userContext, e);
+		}
 		catch (Exception e) {
 			Log.severe("Exception while compiling and rendering article '" + title + "'", e);
 			return getExceptionRendering(userContext, e);
 		}
 	}
 
-	private Article updateArticle(WikiContext wikiContext, String content) throws InterruptedException {
+	private Article updateArticle(WikiContext wikiContext, String content) throws InterruptedException, UpdateNotAllowedException {
 		HttpServletRequest httpRequest = wikiContext.getHttpRequest();
 		if (httpRequest == null) {
 			// When a page is rendered the first time, the request is null.
@@ -279,23 +287,28 @@ public class KnowWEPlugin extends BasicPageFilter implements WikiPlugin,
 			// we can just ignore it.
 			return null;
 		}
+
 		String title = wikiContext.getPage().getName();
-		Article article = Environment.getInstance().getArticle(Environment.DEFAULT_WEB, title);
 
 		String originalText = "";
+		Article article = Environment.getInstance().getArticle(Environment.DEFAULT_WEB, title);
 		if (article != null) {
 			originalText = article.getRootSection().getText();
 		}
 		String parse = UserContextUtil.getParameters(httpRequest).get("parse");
 		boolean fullParse = parse != null && (parse.equals("full") || parse.equals("true"));
 		if (fullParse || !originalText.equals(content)) {
+			if (!Environment.getInstance().getWikiConnector().userCanEditArticle(title, httpRequest)) {
+				throw new UpdateNotAllowedException();
+			}
 			ArticleManager articleManager = Environment.getInstance().getArticleManager(Environment.DEFAULT_WEB);
 			articleManager.open();
 			try {
 				deleteRenamedArticles(title);
 				article = Environment.getInstance().buildAndRegisterArticle(
 						Environment.DEFAULT_WEB, title, content, fullParse);
-			} finally {
+			}
+			finally {
 				articleManager.commit();
 			}
 			Compilers.getCompilerManager(Environment.DEFAULT_WEB).awaitTermination();
@@ -336,9 +349,16 @@ public class KnowWEPlugin extends BasicPageFilter implements WikiPlugin,
 
 	private String getExceptionRendering(UserContext context, Exception e) {
 		RenderResult renderResult = new RenderResult(context.getRequest());
-		String message = "An exception occured while compiling and rendering this article, "
-				+ "try going back to the last working version of the article.\n\n"
-				+ ExceptionUtils.getStackTrace(e);
+		String message;
+		if (e instanceof UpdateNotAllowedException) {
+			message = "Your request tries to change the content of the current article. "
+					+ "You are not authorized to do that.";
+		}
+		else {
+			message = "An exception occurred while compiling and rendering this article, "
+					+ "try going back to the last working version of the article.\n\n"
+					+ ExceptionUtils.getStackTrace(e);
+		}
 		renderResult.appendHtmlElement("div", message, "style",
 				"white-space: pre; overflow: hidden");
 		return renderResult.toStringRaw();
@@ -537,4 +557,6 @@ public class KnowWEPlugin extends BasicPageFilter implements WikiPlugin,
 		return "";
 	}
 
+	private class UpdateNotAllowedException extends Exception {
+	}
 }
